Rewrite as a single file
