+++ sample.py
@@ -1,13 +1,79 @@
 from typedb.driver import TypeDB, SessionType, TransactionType, TypeDBOptions
-import setup
-import argparse
 
-DB_NAME = setup.DB_NAME
-SERVER_ADDR = setup.SERVER_ADDR
+DB_NAME = "sample_app_db"
+SERVER_ADDR = "127.0.0.1:1729"
 
-parser = argparse.ArgumentParser(description='Sample application.')
-parser.add_argument('--reset', '--force', action='store_true', help='Reset the database if it exists.')
-args = parser.parse_args()
+
+def create_new_database(driver, db_name, db_reset=False):
+    if driver.databases.contains(db_name):
+        if db_reset:
+            print("Replacing an existing database", end="...")
+            driver.databases.get(db_name).delete()  # Delete the database if it exists already
+            driver.databases.create(db_name)
+            print("OK")
+            return True
+        else:  # db_reset = False
+            answer = input("Found a pre-existing database. Do you want to replace it? (Y/N) ")
+            answer = answer.lower()
+            if answer == "y":
+                return create_new_database(driver, db_name, db_reset=True)
+            else:
+                print("Reusing an existing database. To reset the database, consider using the --reset argument.")
+                return False
+    else:  # No such database found on the server
+        print("Creating a new database", end="...")
+        driver.databases.create(db_name)
+        print("OK")
+        return True
+
+
+def db_schema_setup(schema_session, schema_file='iam-schema.tql'):
+    with open(schema_file, 'r') as data:
+        define_query = data.read()
+    with schema_session.transaction(TransactionType.WRITE) as tx:
+        print("Defining schema", end="...")
+        tx.query.define(define_query)
+        tx.commit()
+        print("OK")
+
+
+def db_dataset_setup(data_session, data_file='iam-data-single-query.tql'):
+    with open(data_file, 'r') as data:
+        insert_query = data.read()
+    with data_session.transaction(TransactionType.WRITE) as tx:
+        print("Loading data", end="...")
+        tx.query.insert(insert_query)
+        tx.commit()
+        print("OK")
+
+
+def test_initial_database(data_session):
+    with data_session.transaction(TransactionType.READ) as tx:
+        test_query = "match $u isa user; get $u; count;"
+        print("Testing the database", end="...")
+        response = tx.query.get_aggregate(test_query)
+        result = response.resolve().as_value().as_long()
+        if result == 3:
+            print("Passed")
+            return True
+        else:
+            print("Failed with the result:", result, "\n Expected result: 3.")
+            return False
+
+
+def db_setup(driver, db_name, db_reset=False):
+    print(f"Setting up the database: {db_name}")
+    new_database = create_new_database(driver, db_name, db_reset)
+    if not driver.databases.contains(db_name):
+        print("Database creation failed. Terminating...")
+        exit()
+    if new_database:
+        with driver.session(db_name, SessionType.SCHEMA) as session:
+            db_schema_setup(session)
+        with driver.session(db_name, SessionType.DATA) as session:
+            db_dataset_setup(session)
+    with driver.session(db_name, SessionType.DATA) as session:
+        return test_initial_database(session)
 
 
 def fetch_all_users(driver) -> list:
@@ -125,7 +191,7 @@ def delete_file(driver, path):
 
 def main():
     with TypeDB.core_driver(SERVER_ADDR) as driver:
-        if not setup.db_setup(driver, DB_NAME, db_reset=args.reset):
+        if not db_setup(driver, DB_NAME, db_reset=False):
             print("Terminating...")
             exit()
